chore: update docker & flyway configurations
Includes a default `Dockerfile` instead of the `.dev` version, along
with a `.dockerignore` to ignore files/folders for the final build
image.

Updated the `flyway.conf` to declare environment variables used in the
k8s `initContainer` deployment envs.

Added docs on how to install, configure and use `flyway`.

fixes issue #6